pyproject.toml
[project]
name = "azure-search-mcp"
version = "0.1.0"
description = "MCP server for Azure AI Search"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp[cli]>=1.1.2",
    "azure-search-documents>=11.5.2",
    "click>=8.0.0",
]

[project.scripts]
azure-search-mcp = "azure_search_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/azure_search_mcp"]
